requirements.txt
streamlit
pandas
scikit-learn
joblib
python-docx